[Devin] refactor mdxai cli
