+++ packages/mdxai/src/commands/index.ts
@@ -1,0 +1,15 @@
+import { generate } from './generate.js'
+import { edit } from './edit.js'
+import { list } from './list.js'
+import { research } from './research.js'
+import { deepwiki } from './deepwiki.js'
+
+export { generate, edit, list, research, deepwiki }
+
+export const commands = {
+  generate,
+  edit,
+  list,
+  research,
+  deepwiki,
+}
