+++ schema/.mdx/index.d.ts
@@ -1,8 +1,8 @@
 // This file is generated by Velite
 
-import type __vc from '../.velite.temp.js'
+import type { collections } from 'mdxld'
 
-type Collections = typeof __vc.collections
+type Collections = typeof collections
 
 export type mdx = Collections['mdx']['schema']['_output']
 export declare const mdx: mdx[]
